requirements/base.txt
Django>=3.0.4,<4.0
djangorestframework>=3.11.0,<4.0
python-slugify>=5.0.2,<6.0
